requirements.txt
django
graphene_django
graphene

google-auth
google-cloud-bigquery

pandas
db-dtypes